add 'getByDate' generic
