--- logic/agendas.js
@@ -1,25 +1,0 @@
-﻿var ObjectId = require('mongodb').ObjectID;
-var Repository = require('../model/repositoryv2.js');
-require('../helpers/dateHelper.js');
-require('../helpers/numberHelper.js');
-
-var agendasRepository = new Repository('agendas');
-
-exports.getByDate = function (req, res) {
-    var date = new Date(req.params.date);
-    var dateTime = date.getTime();
-    agendasRepository.find({ CreatedBy: req.params.author }).then(function (docs) {
-        var agendas = docs.filter(function (a) {
-            return a.Date >= date.getFirstMs() && a.Date <= date.getLastMs();
-        });
-        var dates = docs.map(function (a) { return a.Date; });
-        var prevs = dates.filter(function (d) { return d < dateTime; }).sort(function (d1, d2) { return d2 - d1; });
-        var nexts = dates.filter(function (d) { return d > dateTime; }).sort(function (d1, d2) { return d1 - d2; });
-        var results = {
-            Prev: prevs.length == 0 ? null : new Date(prevs[0]).toAnyString(),
-            Next: nexts.length == 0 ? null : new Date(nexts[0]).toAnyString(),
-            Agendas: agendas
-        };
-        res.send(results);
-    });
-}
