+++ logic/genericLogicv2.js
@@ -43,6 +43,26 @@ exports.insertv2 = function (req, res) {
     });
 }
 
+exports.getByDate = function (req, res) {
+    var date = new Date(req.params.date);
+    var dateTime = date.getTime();
+    var repository = new Repository(req.params.collectionName);
+    repository.find({ CreatedBy: req.params.author }).then(function (docs) {
+        var records = docs.filter(function (a) {
+            return a.Date >= date.getFirstMs() && a.Date <= date.getLastMs();
+        });
+        var dates = docs.map(function (a) { return a.Date; });
+        var prevs = dates.filter(function (d) { return d < dateTime; }).sort(function (d1, d2) { return d2 - d1; });
+        var nexts = dates.filter(function (d) { return d > dateTime; }).sort(function (d1, d2) { return d1 - d2; });
+        var results = {
+            Prev: prevs.length == 0 ? null : new Date(prevs[0]).toAnyString(),
+            Next: nexts.length == 0 ? null : new Date(nexts[0]).toAnyString(),
+            Docs: records
+        };
+        res.send(results);
+    });
+}
+
 exports.searchv2 = function (req, res) {
     var repository = new Repository(req.params.collectionName);
     var columns = req.body.filters;
